requirements.txt
asgiref==3.8.1
attrs==23.2.0
black==24.4.2
click==8.1.7
colorama==0.4.6
Django==4.0.4
django-debug-toolbar==3.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pep8-naming==0.13.2
pillow==10.3.0
platformdirs==4.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rest-framework-simplejwt==0.0.2
rpds-py==0.18.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
